Move the executor template specializations
We move the template specializations related to the executors.
Change the future.hpp include to the future_exec.hpp one when
necessary.
